Fix sticky collisions in simulation loop
Outdated distance matrix was used to decide whether particles collided. To do: rearrange computations so that distance does not need to be calculated twice (as this is the most expensive step).
Note that too small epsilon and collision threshold caused overshooting, which resembles non-sticky collisions

In 1D, additionally improve plotting